Fix API page limit error
This is an error that made it impossible for the script to be "infinite" based on the fact that the AL API only allows 50 elements per page, and I didn't have a pagination system.
I made a really rough pagination system that will break at the touch of a feather, but it works if used as intended.

It will make a cycle of 30 posts, wait 60 seconds for the API rate limit to refresh and then go for another page.

Will probably fix it in the future, tho!
Changes to be committed:
    * modified:   main.py