Edit main.py flow
This is a combination of 2 commits.
This is the 1st commit message:

Add error handling 429 back

Until I figure a proper solution to the whole script, I believe this should be added back, just so that the script doesn't die out inmediately if the API rate limit is hit before expected.

This is the commit message #2:

It makes little sense to make the user wait 60 seconds just for the script to close. This commit should fix that issue.

This is untested, I don't think it makes anything break.
Changes to be committed:
    * modified:   Pipfile.lock
    * modified:   main.py